+++ rollup.npm.js
@@ -85,10 +85,7 @@ export default [
     input: 'packages/main.js',
     external: ['react'],
     globals: { react: 'React' },
-    output: [
-      { file: pkg.unpkg, format: 'umd' },
-      { file: 'docs/react/boardgameio.min.js', format: 'umd' },
-    ],
+    output: [{ file: pkg.unpkg, format: 'umd' }],
     name: 'BoardgameIO',
     plugins: plugins.concat([
       replace({